src/users/users.service.spec.ts
import 'reflect-metadata';
import { Container } from 'inversify';
import { IConfigService } from '../config/config.service.interface';
import { IUsersRepository } from './users.repository.interface';
import { IUsersService } from './users.service.interface';
import { TYPES } from '../types';
import { UsersService } from './users.service';
import { UserModel } from '@prisma/client';


const ConfigServiceMock: IConfigService = {
	get: jest.fn(),
};

const UsersRepositoryMock: IUsersRepository = {
	find: jest.fn(),
	create: jest.fn(),
};

const container = new Container();
let configService: IConfigService;
let usersRepository: IUsersRepository;
let usersService: IUsersService;


beforeAll(() => {
	container.bind<IUsersService>(TYPES.UserService).to(UsersService);
	container.bind<IConfigService>(TYPES.ConfigService).toConstantValue(ConfigServiceMock);
	container.bind<IUsersRepository>(TYPES.UsersRepository).toConstantValue(UsersRepositoryMock);


	configService = container.get<IConfigService>(TYPES.ConfigService);
	usersRepository = container.get<IUsersRepository>(TYPES.UsersRepository);
	usersService = container.get<IUsersService>(TYPES.UserService);
});


describe('User Service', () => {

	let createdUser: UserModel | null  = null


	it('create User', async () => {
		configService.get = jest.fn().mockReturnValueOnce('1');
		usersRepository.create = jest.fn().mockImplementationOnce((user: UserModel) => {
			return ({
				name: user.name,
				email: user.email,
				fio: user.fio,
				password: user.password,
				id: 1,
			});
		});
		createdUser = await usersService.createUser({
			email: 'aaa@aaa.ru',
			fio: 'fio',
			name: 'name',
			password: '123123',
		});

		expect(createdUser?.id).toEqual(1);
		expect(createdUser?.password).not.toEqual('123123');

	});
	it('validate user success', async () => {
		usersRepository.find = jest.fn().mockReturnValueOnce(createdUser);

		const res = await usersService.validateUser({ email: 'aaa@aaa.ru', password: '123123' });

		console.log(res);

		expect(res).toBeTruthy()
	});


});